Refactored datasource from Main webserver class into its own DatasourceManagerclass
